+++ src/lib/net/encryption/src/lib.rs
@@ -1,5 +1,5 @@
 pub mod errors;
-pub mod key;
+pub mod rsa_enc;
 pub fn add(left: u64, right: u64) -> u64 {
     left + right
 }
